pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Fuckdl"
version = "0.1.0"
description = "Playready and Widevine DRM downloader and decrypter with multi-VPN proxy support"
authors = []

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

# Core dependencies
appdirs = "^1.4.4"
click = "^8.1.3"
coloredlogs = "^15.0"
PyYAML = "^6.0.1"
tqdm = "^4.67.1"
requests = { extras = ["socks"], version = "^2.32.3" }
httpx = "^0.27.0"
urllib3 = "^2.2.0"

setuptools = ">=70.0.0,<81.0.0"

# DRM and decryption
cryptography = "^43.0.3"
pycryptodome = "^3.21.0"
pycryptodomex = "^3.21.0"
pywidevine = "1.8.0"
pyplayready = { path = "./scripts/pyplayready-0.6.0", develop = true }

# Network and proxies
aiohttp = "^3.9.1"
curl-cffi = "^0.14.0"
pproxy = "^2.7.7"
pysocks = "^1.7.1"
tenacity = "^8.2.3"
backoff = "^2.2.1"
async-timeout = "^4.0.3"

# Video/Media processing
m3u8 = "^3.0.0"
lxml = "^5.3.0"
beautifulsoup4 = "^4.12.0"
pymediainfo = "^6.1.0"
pycaption = "^2.1.1"
pysubs2 = "^1.6.1"
webvtt-py = "^0.5.1"
xmltodict = "^0.14.2"

# Subtitles and language
langcodes = { extras = ["data"], version = "^3.4.0" }
language-data = ">=1.3.0,<2.0.0" 
pycountry = "^24.6.1"
Unidecode = "^1.2.0"

# Utilities
construct = "2.8.8"
crccheck = "^1.0"
ecpy = "^1.2.5"
isodate = "^0.6.1"
jsonpickle = "^3.0.0"
marisa-trie = "^1.1.0"
protobuf = ">=4.21.0,<5.0.0"
PyJWT = "^2.8.0"
PyMySQL = { extras = ["rsa"], version = "^1.0.2" }
ruamel-yaml = "^0.18.10"
tldextract = "^5.1.0"
toml = "^0.10.2"
validators = "^0.34.0"
websocket-client = "^1.8.0"
wasmtime = "^24.0.0"
rich = "^15.0.0"
sortedcontainers= "^2.4.0"

# Downloaders
yt-dlp = "^2026.3.17"

# Docker support for Gluetun
docker = { version = "^7.0.0", optional = true }

# Subby
subby = { path = "./scripts/subby", develop = true }

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.0.0"
flake8 = "^7.0.0"
isort = "^5.13.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
mypy = "^1.8.0"

[tool.poetry.group.vpn]
optional = true
dependencies = { docker = "^7.0.0" }

[tool.poetry.scripts]
fuckdl = "fuckdl.fuckdl:main"
[tool.poetry.extras]
vpn = ["docker"]

[tool.isort]
line_length = 120
classes = ["CTV", "FPS", "IO", "iTunes", "MP4", "TVNOW"]
extend_skip = ["fuckdl/vendor"]
profile = "black"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true